refactor(checkbox): rewrite component
