+++ source/components/atoms/Checkbox/Checkbox.tsx
@@ -1,56 +1,18 @@
 import React, { useContext } from "react";
-import PropTypes from "prop-types";
-import styled, { ThemeContext } from "styled-components/native";
-import type { PrimaryColor } from "../../../theme/themeHelpers";
+import { ThemeContext } from "styled-components/native";
+
 import { getValidColorSchema } from "../../../theme/themeHelpers";
-import Icon from "../Icon";
 
-interface BoxProps {
-  checked: boolean;
-  colorSchema: PrimaryColor;
-  backgroundColor: string;
-  size: "small" | "medium" | "large";
-}
-const CheckboxBox = styled.TouchableHighlight<BoxProps>`
-  border-style: solid;
-  border-color: ${(props) =>
-    props.checked
-      ? "transparent"
-      : props.theme.colors.complementary[props.colorSchema][0]};
-  background-color: ${({ checked, backgroundColor }) =>
-    checked ? backgroundColor : "transparent"};
-  width: ${(props) => props.theme.checkbox[props.size].width}px;
-  height: ${(props) => props.theme.checkbox[props.size].height}px;
-  padding: ${(props) => props.theme.checkbox[props.size].padding}px;
-  margin: ${(props) => props.theme.checkbox[props.size].margin}px;
-  border-width: ${(props) => props.theme.checkbox[props.size].borderWidth}px;
-  border-radius: ${(props) => props.theme.checkbox[props.size].borderRadius}px;
-`;
+import { CheckboxBox, CheckboxTick } from "./Checkbox.styled";
 
-interface CheckboxProps {
-  color: string;
-}
-const CheckboxTick = styled(Icon)<CheckboxProps>`
-  color: ${({ color }) => color};
-  margin-left: -2px;
-  margin-top: -3px;
-`;
-
-interface Props {
-  checked?: boolean;
-  onChange: () => void;
-  disabled?: boolean;
-  invertColors?: boolean;
-  size?: "small" | "medium" | "large";
-  colorSchema: PrimaryColor;
-}
+import type { Props } from "./Checkbox.types";
 
 const Checkbox: React.FC<Props> = ({
   checked,
   onChange,
-  size,
-  disabled,
-  colorSchema,
+  size = "small",
+  disabled = false,
+  colorSchema = "blue",
   invertColors = false,
 }) => {
   const { colors, checkbox } = useContext(ThemeContext);
@@ -79,7 +41,7 @@ const Checkbox: React.FC<Props> = ({
       checked={checked}
       colorSchema={validColorSchema}
       underlayColor={underlayColor}
-      size={size || "small"}
+      size={size}
       backgroundColor={checkboxColor}
     >
       {checked ? (
@@ -93,41 +55,6 @@ const Checkbox: React.FC<Props> = ({
       )}
     </CheckboxBox>
   );
-};
-
-Checkbox.propTypes = {
-  /**
-   * Boolean that determines if the checkbox is checked or not. Manages the 'state' of the component.
-   */
-  checked: PropTypes.bool.isRequired,
-  /**
-   * What happens when the checkbox is clicked.
-   */
-  onChange: PropTypes.func,
-  /**
-   * sets the color theme.
-   */
-  colorSchema: PropTypes.oneOf(["red", "blue", "green", "purple", "neutral"]),
-  /**
-   * One of small, medium, large
-   */
-  size: PropTypes.oneOf(["small", "medium", "large"]),
-  /**
-   * Disables the checkbox if true.
-   */
-  disabled: PropTypes.bool,
-  /**
-   * Make only the tick icon visible when inverting colors.
-   */
-  invertColors: PropTypes.bool,
-};
-
-Checkbox.defaultProps = {
-  onChange: () => undefined,
-  colorSchema: "blue",
-  size: "small",
-  disabled: false,
-  invertColors: false,
 };
 
 export default Checkbox;
